requirements.txt
# AI Resume Intelligence Analyzer — Dependencies
# ================================================

# Core ML / Numeric
numpy>=1.24.0
scikit-learn>=1.3.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0

# HTTP / Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# PDF Parsing (pdfminer.six is primary, PyPDF2 is fallback)
pdfminer.six>=20221105
PyPDF2>=3.0.0

# Groq AI
groq>=0.9.0

# Environment variable loading
python-dotenv>=1.0.0

# Web API
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6